+++ bywaf/__init__.py
@@ -1,3 +1,7 @@
 """Bywaf package."""
 
 __version__ = "0.1.0"
+
+from .api import BywafSession
+
+__all__ = ["BywafSession", "__version__"]